Move remaining tests for writing to workflow_orchestrator_test.py
