+++ test/unit/util/workflows/workflow_test.py
@@ -5,21 +5,15 @@ from unittest.mock import patch
 import pytest
 
 from exasol.toolbox.util.workflows.exceptions import (
-    InvalidWorkflowPatcherEntryError,
     TemplateRenderingError,
-    YamlJobValueError,
     YamlOutputError,
     YamlParsingError,
 )
 from exasol.toolbox.util.workflows.process_template import WorkflowRenderer
 from exasol.toolbox.util.workflows.templates import (
-    NOT_MAINTAINED_WORKFLOW_NAMES,
     WORKFLOW_TEMPLATE_OPTIONS,
 )
-from exasol.toolbox.util.workflows.workflow import (
-    Workflow,
-    update_workflow,
-)
+from exasol.toolbox.util.workflows.workflow import Workflow
 
 
 @pytest.fixture
@@ -229,151 +223,3 @@ class TestWorkflow:
                     output_directory=tmp_path,
                     github_template_dict=project_config.github_template_dict,
                 )
-
-
-class TestUpdateWorkflow:
-    @staticmethod
-    def test_works_as_expected_without_patcher(project_config_without_patcher):
-        workflow_name = "merge-gate"
-        # setup
-        project_config_without_patcher.github_workflow_directory.mkdir(parents=True)
-        input_text = WORKFLOW_TEMPLATE_OPTIONS[workflow_name].read_text()
-        expected_file_path = (
-            project_config_without_patcher.github_workflow_directory
-            / f"{workflow_name}.yml"
-        )
-
-        update_workflow(
-            workflow_choice=workflow_name, config=project_config_without_patcher
-        )
-        result = expected_file_path.read_text()
-
-        # Currently, we check only a subselection as we must preserve formatting for tbx
-        # endpoints, and there are 2 minor whitespace differences.
-        assert result[:10] == input_text[:10]
-
-    @staticmethod
-    def test_works_as_expected_with_relevant_patcher(project_config, remove_job_yaml):
-        # remove_job_yaml modifies "checks" and that's also the workflow being updated
-        workflow_name = "checks"
-        # setup
-        project_config.github_workflow_directory.mkdir(parents=True)
-        input_text = WORKFLOW_TEMPLATE_OPTIONS[workflow_name].read_text()
-        expected_file_path = (
-            project_config.github_workflow_directory / f"{workflow_name}.yml"
-        )
-        # setup checks
-        removed_job_name = "build-documentation-and-check-links"
-        assert removed_job_name in remove_job_yaml
-        assert removed_job_name in input_text
-
-        update_workflow(workflow_choice="checks", config=project_config)
-        result = expected_file_path.read_text()
-
-        # We compare only a subselection to verify that the files are roughly the
-        # same, and we expect them to differ as the 'result' does not contain
-        # the 'removed_job_name'
-        assert result[:10] == input_text[:10]
-        assert removed_job_name not in result
-
-    @staticmethod
-    def test_works_as_expected_with_not_relevant_patcher(
-        project_config, remove_job_yaml
-    ):
-        # remove_job_yaml modifies "checks" and that's NOT the workflow being updated
-        workflow_name = "merge-gate"
-        # setup
-        project_config.github_workflow_directory.mkdir(parents=True)
-        input_text = WORKFLOW_TEMPLATE_OPTIONS[workflow_name].read_text()
-        expected_file_path = (
-            project_config.github_workflow_directory / f"{workflow_name}.yml"
-        )
-
-        update_workflow(workflow_choice=workflow_name, config=project_config)
-        result = expected_file_path.read_text()
-
-        # Currently, we check only a subselection as we must preserve formatting for tbx
-        # endpoints, and there are 2 minor whitespace differences.
-        assert result[:10] == input_text[:10]
-
-    @staticmethod
-    def test_not_maintained_workflows_added_to_new_project(
-        project_config_without_patcher,
-    ):
-        directory = project_config_without_patcher.github_workflow_directory
-        directory.mkdir(parents=True)
-
-        update_workflow(workflow_choice="all", config=project_config_without_patcher)
-
-        assert all(
-            (directory / f"{name}.yml").exists()
-            for name in NOT_MAINTAINED_WORKFLOW_NAMES
-        )
-
-    @staticmethod
-    @pytest.mark.parametrize("workflow_name", NOT_MAINTAINED_WORKFLOW_NAMES)
-    def test_not_maintained_workflows_not_modified_in_old_project(
-        project_config_without_patcher, workflow_name
-    ):
-        directory = project_config_without_patcher.github_workflow_directory
-        directory.mkdir(parents=True, exist_ok=True)
-        workflow = "slow-checks.yml"
-        (directory / workflow).touch()
-
-        update_workflow(
-            workflow_choice=workflow_name, config=project_config_without_patcher
-        )
-
-        assert {file_path.name for file_path in directory.iterdir()} == {workflow}
-        assert (directory / workflow).read_text() == ""
-
-    @staticmethod
-    @pytest.mark.parametrize("workflow_name", NOT_MAINTAINED_WORKFLOW_NAMES)
-    def test_not_maintained_workflows_not_added_to_old_project(
-        project_config_without_patcher, workflow_name
-    ):
-        directory = project_config_without_patcher.github_workflow_directory
-        directory.mkdir(parents=True, exist_ok=True)
-        (directory / "dummy.yml").touch()
-
-        update_workflow(
-            workflow_choice=workflow_name, config=project_config_without_patcher
-        )
-
-        assert {file_path.name for file_path in directory.iterdir()} == {"dummy.yml"}
-
-    @staticmethod
-    @pytest.mark.parametrize("workflow_name", NOT_MAINTAINED_WORKFLOW_NAMES)
-    def test_not_maintained_workflows_not_modified_in_old_project(
-        project_config_without_patcher, workflow_name
-    ):
-        directory = project_config_without_patcher.github_workflow_directory
-        directory.mkdir(parents=True, exist_ok=True)
-        workflow = "slow-checks.yml"
-        (directory / workflow).touch()
-
-        update_workflow(
-            workflow_choice=workflow_name, config=project_config_without_patcher
-        )
-
-        assert {file_path.name for file_path in directory.iterdir()} == {workflow}
-        assert (directory / workflow).read_text() == ""
-
-    @staticmethod
-    def test_raises_invalidworkflowpatcherentryerror(project_config):
-        patcher_yml = """
-        workflows:
-        - name: "checks"
-          remove_jobs:
-            - unknown-job
-        """
-        project_config.github_workflow_patcher_yaml.write_text(patcher_yml)
-
-        with pytest.raises(InvalidWorkflowPatcherEntryError) as ex:
-            update_workflow(workflow_choice="checks", config=project_config)
-
-        assert (
-            f"In file '{project_config.github_workflow_patcher_yaml}', "
-            "an entry '{'job_name': 'unknown-job'}' does not exist in"
-        ) in str(ex.value)
-        assert isinstance(ex.value.__cause__, YamlJobValueError)
